backend/requirements.txt
django
djangorestframework
django-cors-headers
firebase-admin
